Split networking functions out
